Update murano service name
In Mitaka, the service name must have a dash rather than underscore when using
the sql catalog driver in Keystone[0] (the default).

This works for upgrade also, though further improvements could be written to
remove the old endpoint from Keystone, and automatically chose a dash or
underscore based on driver type used.

[0] http://docs.openstack.org/releasenotes/murano/mitaka.html#upgrade-notes

Change-Id: I15a03370afdad6529eec51a206b6134bf80b283d
Closes-Bug: 1576152