Remove dependencies on kazoo and friends
Since kolla-mesos has been retired, there is no need in the code
base to keep the zookeeper implementation.  As a result, just remove
it.  If we were to keep it we need virtualenvs for that part of the
code base which installs the python dependencies related to it
for from-binary installs.

This just simplifies the implementation tremendously and culls dead
code.

Closes-Bug: #1577194
(cherry picked from commit 025d57f82005a4ea92950f9b4d717c88f647d15f)

Conflicts:
	docker/base/Dockerfile.j2
	docker/kolla-toolbox/Dockerfile.j2

Change-Id: Ieda226e652d67f5b5667112f4f2556f3171366d3