Avoid _ in rabbitmq_clusterer.config
Trying to use ConfigMap's in Kubernetes leads to an interesting
problem. We use the file name as the key and the contents of the
file as the text value. The ConfigMap is mounted on the container
as a volume and the key is then used as the name of the file. The
problem is that kubernetes has a limitation on the name of the
key

https://github.com/kubernetes/kubernetes/blob/release-1.2/docs/design/identifiers.md

Which means we cannot use '_' in the name of the file.

Closes-Bug: #1581162
Change-Id: I2d9ec80f989c30893b019954fe18b3623d27a076