Generate configuration using ansible for any project
Give outside projects the ability to consume kolla-ansible as a
config generation mechanism by allowing for the developer to define
a directory structure within /etc/kolla that fits a project's needs.

For example, kolla-kubernetes does not define configuration in terms
of serivces, but in terms of pods (groups of services).

Partially implements: blueprint multi-project-config

Change-Id: If08443aefcaf5fa5604b243b39a6a2d54e431cff