Fix the rabbitmq crash on CentOS binary and source deploy
this is a wordround fix for the rabbitmq failed when deploy on CentOS in
the CI gate. the ideal fix should set the hostname in setup_gate.sh
script. But it do not work as expect with unknown reason

Partial-Bug: #1581301
Change-Id: Ia692794aadf11d3ebe75e4e2bf6b42d9a5f3996f
(cherry picked from commit c98e9c44f23d14867e4dddea88fe9a9f7aa50dd6)