setup_{RedHat,Debian}.sh, make more idempotent
Previously, when either setup_{RedHat,Debian}.sh was ran
multiple times a mess was created due to multiple loop
devices being setup.

This change now looks to see if '/docker' or '/swapfile'
already exist, and not re-create them if they do.

*However*, /var/lib/docker (/docker) is still wiped - but it
means that the script can run and bring the system into a
constant vanilla state.

Change-Id: I169662fa04c2bf644672ca75c3ee579795943727
Closes-Bug: #1586563
Signed-off-by: Dave Walker (Daviey) <email@daviey.com>