Revert "Added precheck to check kernel version 4.2 on Ubuntu 14.04."
Upgrade the kernel version is not acceptable for the server in most
of case. I do not see the issue when using ubuntu 14.04 default
kernel with docker aufs storage driver.

Even if there is some wrong in the aufs driver, we could recommend
the user to use other drivers, like lvm, brtfs or even overlay driver.

This reverts commit 4c6e4676fcef4696c70c5813f65b3fa96550e6d4.

Change-Id: I5be2971eb1d2021bbc1980c4279dc1181a7da23f