Revert "Added precheck to check kernel version 4.2 on Ubuntu 14.04."
Upgrade the kernel version is not acceptable for the server in most
of case. I do not see the issue when using ubuntu 14.04 default
kernel with docker aufs storage driver.

Even if there is some wrong in the aufs driver, we could recommend
the user to use other drivers, like lvm, brtfs or even overlay driver.

This reverts commit 833c1269c7941498787d49ce879be6c283728407.

Change-Id: I6ee71ecd1ca64777b874bfcc342162622643b89c