glance-*, fix rabbit config with ceilometer
Currently, if ceilometer is enabled with glance then the
rabbitmq is non-functional, causing various glance failures
including image upload.

This is caused because:
The generated config fails to have a carriage return at the
end of the rabbit_host line, causing the following line to
be added to this entry.

rabbit_host is used rather than rabbit_hosts, where the
fromer expects a string, rather than the list that is
provided.

This also adds rabbit HA support, to be in-line with the
other services using rabbitmq.

This is resolved, by pushing the rabbit_hosts entry to be
last, meaning that a carriage return is not required.

Change-Id: I89f26d542565b98025fd9b84fc4beb5cbed364e5
Closes-Bug: 1596244
Signed-off-by: Dave Walker (Daviey) <email@daviey.com>