Build docker images for OpenStack Watcher
Watcher is part of the OpenStack big-tent and is formally
known as "OpenStack Infrastructure Optimization service".

Whilst it provides a range of default goals and strategies,
the most relevant case is to enable re-balancing of the
compute hosts by taking CPU usage (data from Ceilometer)
into account and live migrating instances as required.

Currently this only builds the Docker images for type source
but binary builds are gracefully ignored.

The ansible configurations will be part of a later commit.

Change-Id: I9bb81ee625d9fcf6513e44e2ed20384e34da2adc
Partial-bug: #1598929
Partially-implements: bp watcher
Signed-off-by: Dave Walker (Daviey) <email@daviey.com>