Add guidelines on adding a new service
Add docs to provide more info and guidelines on how to add a new service
to Kolla.

There are a lot of patterns and practices that have evolved in Kolla so
it's good to formalise them to make it easier for both developers and
reviewers to ensure nothing has being missed.

Change-Id: I9ab858fdd0c756e2505fc0e3487bd8e026da5db1