OpenStack Services Should Use keystone_internal_url for auth
Horizon and Neutron mistakenly were using keystone_public_url
for authentication.  This works without error in deployments
when the internal services happen to have access to the
public network, but it is still wrong.  This fails to work
when the internal services can not access the public URLs,
for example when TLS is enabled on the public endppoints.

This patches corrects horizon and neutron to use
keystone_internal_url for auth.

Change-Id: I59b9094364bef375036028ba86a771dabf28c963
Closes-bug: #1625648