Add http as allowed store when glance configured to use rbd
Adds http as an allowed store when glance is configured to use
ceph/rbd as the image store. The default configuration for
glance with lvm uses http as a secondary store. http store
is commonly used with automation tools like heat to pull
images into glance from endpoints. Without specifying http
as a store the expected default behavior fails. This patch
restores that default behavior.

Change-Id: Id3663123b8d54e53bd1fd130b65fd8383c9f35c0
Closes-Bug: #1631637