Fix plugin parsing for base images
If we try to add a plugin to a base image, e.g.

[neutron-base-plugin-neutron-fwaas]
type = git
location = https://github.com/openstack/neutron-fwaas
reference = master

The build will fail as it parses the above as both
"neutron-base-plugin-neutron-fwaas" and "base-plugin-neutron-fwaas".

Update the regex to ensure the image name is anchored to the beginning of the
conf section.

Change-Id: I4099ebd0d25cd28a9bb3bdc4fcfbf311ce22a3fd
Closes-Bug: #1638900