Move manila-share service to network node
After a research regarding manila-share node,
is required that for the generic driver the
manila-share service should be run on the network node.

As the generic driver is our default driver, so this change is needed.

Closes-Bug: #1639028
Change-Id: I4b73fec4f96e3e2f9c442c8716941bede5dfbc8e
(cherry picked from commit a2707ecf0043d5df25d0f87d6dde55aeb52260da)