Add a 'fallback mode' to disable udev in find_disks
As part of change I362b3f8e91de79687fc84e256996fbcaf303b6af a check was
added to use udev to read GPT labels only if the version of udev
installed supports that, otherwise fallback to sgdisk.

Unfortunately the check for udev version is unreliable, as it is
returning the version of the udev userspace tools installed within the
container rather than on the host.

As well as this, we're seeing the find_disks.py having trouble reading
the UUID of a partition via udev. The test system is Oracle VM 3.4.2
which was released earlier this year, but has an old udev (derivative of
RHEL 6).

Even if we use Ansible to query the host's version of udev and pass that
through to the container, querying the above disk attributes is a
lottery on older systems without a reliable reference of which versions
of udev support which operations. Hence we should give operators a
reliable way to toggle it's usage.

Partial-Bug: 1631949
Change-Id: I6ad7825cdb164498f3d02f2ae064c7c1c38e10d5