Not to pull the base images when building.
Add --pull/--nopull parameter for kolla-build, when --nopull is
given and the base image is already loaded, The docker-engine will
not attemp to pull a newer version of the image. This makes sense
when the registry server of base images is unreachable.

This feature is as same to setting "image_pull_policy" to
"IfNotPresent" while building images.

The reason why I create this feature is that, even if I have setted
base_distro_tag to a stable tag, e.g. "7" or "2016-12-13", docker
still have to wait much time to fetch the version, and return failure
when registry server is unreachable. This is a feature of docker
build --pull=True.

Change-Id: If1bd08763fb757b4f090a3163dd1dbfb8fc531d3
Implements: blueprint docker-build-nopull-base