Removed percona from ubuntu repos
16.04 provides the percona-xtrabackup scripts in the main repos, so the
specific percona repo can be removed. Unfortunately, mariadb10.0 does
not build with galera built-in. So until ubuntu packages mariadb10.1,
we will be dependant on the external repos for mariadb with galera
built in.

Labels keys so they are easily identified in the future. In Jinja2 the
comments cannot go inline while generating the list.

Change-Id: I5e6c2f884c8d59bbc574b71ac731d1b4db18b68c