Add Apache packages to cinder-api container
Some projects that use Kolla docker containers (TripleO) run cinder-api under
httpd. This patch adds the httpd package so that it exists in the container
and can optionally by used via Docker entrypoints.

As many of these projects do not require the use of Apache (it is opt-in) at
this point not all deployment frameworks support the use of it so
the existing configuration defaults have been left as-is for now.

Partial-Blueprint: apache-packages-for-apis

Change-Id: Ibf9665c268465971f92218833542e20623d1d296