Add Apache packages to ironic-pxe container
Some projects that use Kolla docker containers (TripleO) run Ironic pxe under
httpd. This patch adds the httpd package so that it exists in the container
and can optionally by used via Docker entrypoints.

As many of these projects do not require the use of Apache (it is opt-in) at
this point not all deployment frameworks support the use of it so
the existing configuration defaults have been left as-is for now.

Change-Id: I4c3e44db6a1ffa7238006c27038ac0a67c8aa725
Partial-Blueprint: apache-packages-for-apis