Code coverage improvement for kolla
This patch add the coverage tool package to test-requirements.txt
which was missing, while kolla support coverage. In addtion to this
tox.ini coverage command is modified to reflect the coverage report
in standard output whenever "tox -e cover" command run for better
representation of coverage report.

Change-Id: I7f77c8d5320a7d180586ce1e8773a4f294846331
Partially-Implements: blueprint coverage-increment-for-kolla