Remove all kolla-ansible related docs
The goal of this patch is to reduce the docs hosted for this repo down
to just the container images.

Hyperlinks are provided to each of the available deployment mechanisms,
currently kolla-ansible and kolla-kubernetes.

This is to cut down confusion on which doc patches go where, as well as
the massive duplication we currently have between kolla and
kolla-ansible documentation. Further improvements could involve breaking
out the 'Building Container Images' section into separate pages, e.g.
quickstart, customisation, advanced, etc.

Change-Id: Ia7959c71a241b9d7139ec50b34e36b24fbcf76e6