Fix pushing images to local registry
Adding -e REGISTRY_PROXY_REMOTEURL=https://registry-1.docker.io to the
docker run command in the start-registry script configures the local
registry as a pull through cache. However, pushing of images to the
local registry fails.

Pushing to a registry configured as a pull-through cache is unsupported.
For more information, reference the proxy section in Docker
Documentation (https://docs.docker.com/registry/configuration/)

Corresponding change to the documentation has been made to the
multinode.rst file in the Kolla-ansible repository
https://review.openstack.org/#/c/426858/

Change-Id: I636912f9958490bf751bfa0f449998c61d273241
Closes-Bug: #1659738