Fix binary ironic-inspector rootwrap configuration
When built with the binary install type (at least on CentOS), Ironic
inspector fails to start and the container remains in a restarting
state. The log file shows that it is failing to execute iptables, and
analysis found that this was due to an incorrect rootwrap
configuration. The RDO ironic inspector RPM expects Ironic inspector
to be run as the ironic-inspector user, however Kolla uses the ironic
user. This means that neither of the packaged ironic nor
ironic-inspector sudoers configuration files works for us.

Kolla currently installs a sudoers file pointing to the rootwrap
script in the virtualenv of the source install, but of course this
only makes sense for source installs, and should not be installed for
binary installs.

This change adds a second sudoers file that will work for the binary
install type, and installs the correct sudoers file for the install
type.

Change-Id: I8ecd0b658b8df8f38ddf717fa9443d4dc2896984
Closes-Bug: #1624457