Improve kolla image build script for finding parent images
This change updates the image script to search for the parent image via the
Docker 'FROM' tag throughout the Dockerfile. The script previously checked
for parent name of an image, assuming that first line of Dockerfile always
starts with FROM, but that is not always the case when using custom
Dockerfiles.

Change-Id: I176d99f4e4896b64435e495526154c8f4c684839
Closes-Bug: #1656206