Allow full customisation of kube image binaries
The location of the binaries in these images can be overridden by
setting ENV in the header blocks, but if an operator wants to use a
different install method all together (e.g. rpm), these blocks will
allow that.

Also import the macros in kube-base to allow their use in overrides.

Change-Id: I3e52e315f0103d8a1c95d4489fa3841900fdff75