Workaround for setuptools bug affecting Bifrost
bifrost-3.0.0 is affected by a setuptools bug preventing it from being
installed. This change patches Bifrost with the workaround committed
to the Bifrost master branch in https://review.openstack.org/#/c/437010/.
If/when that change gets backported to Bifrost stable/ocata and then
released as 3.0.1, we can remove this workaround.

Change-Id: I257abd86599304ef1992d10da20130c9b7c25e46
Closes-Bug: #1667308