Update oraclelinux tag to new slim base image
Oracle have released a slimmed down version of their base image, using
the tag '7-slim'[0]. This is over half the size of the standard image.

Note, right now the default base tags are also used to check that we're
building on a supported distro version, e.g. oraclelinux:7.  This breaks
when we want to set a default base tag that is different to what's in
/etc/system-release or /etc/lsb-release, e.g. 7-slim.  Decouple this
info so we have access to both sets of data.

[0] https://blogs.oracle.com/developers/put-your-containers-on-a-diet-with-oracle-linux

Depends-On: Id7df645284f260cd615188db2efe889c9cf8115d
Change-Id: I584ac9514950c541005bb900c19fc1c6c7ed2a5d