Defer handling of /var/lib/neutron/kolla to kolla API
neutron-metadata-agent image expected the /var/lib/neutron/kolla
directory to be present but it was not created as part of the image
building. As a result, the extend_start.sh script fails with:

    stat: cannot stat '/var/lib/neutron/kolla': No such file or directory

This commit defers setting the right ownership for files and directories
to the kolla config API via a config.json file.

Kolla-ansible already mounts the directory as the
neutron_metadata_socket named volume and uses kolla API to set the
ownership and permissions [1].

[1] https://github.com/openstack/kolla-ansible/blob/master/ansible/roles/neutron/templates/neutron-metadata-agent.json.j2

Closes-Bug: #1669487
Change-Id: Ib521b0fc69ab6a72295a2fde3309a001bc3063c2