Add mkfs.vfat to nova-compute source images
Currently Windows and non-declared os_type images (defaults
to Windows) attempt to format ephemeral volumes with
mkfs.vfat but fail as this tool is not available in the
container.

This is resolved by installing dosfstools on the Source
builds (as I believe this is handled in binary builds
through dependencies).

Change-Id: Iefa5fb0a16eed30aea935268f2a9becb1481b018
Closes-bug: #1672199
Signed-off-by: Dave Walker (Daviey) <email@daviey.com>