Add Apache packages to heat containers
Some projects that use Kolla docker containers (TripleO) run heat under
httpd. This patch adds the httpd package so that it exists in the
container and can optionally by used via Docker entrypoints.

As many of these projects do not require the use of Apache (it is
opt-in) at this point not all deployment frameworks support the use of
it so the existing configuration defaults have been left as-is for now.

Change-Id: I4d648c4ce4f5f6ba311bfbbf9c78e06104163c61
Partially-Implements: blueprint apache-packages-for-apis