Refactor and enable shell access
When installing from binary, packaging tools create a user for
Keystone which does not have a shell.  This results in fernet
token operations failing due to the fact that `rsync` cannot
connect over SSH because the Keystone user cannot login.

This patch addresses this issue by enabling shell access
to the Keystone user.

Change-Id: Ic753a94311c5d7d72a82d91c1279dfe982e9afbd
Related-Bug: #1617902