Add HAproxy to source installation packages
The latest release of OpenStack has moved a lot of the neutron
metadata proxy processes to use HAproxy instead.  This package
is not installed by default and causes builds to fail.

This patch addresses this issue by adding the HAproxy package
which should clear out gate problems for source installs.

Change-Id: I2e42828b1ccdb3a79b95bee85648d55508c0f2f9