Update Docker to 1.13.1
There were a few issues blocking Kolla from going up to the
latest version of Docker which is 1.13.  Most of these issues
seem to have been addressed by 1.13.1.

This patch updates the version installed by Kolla to 1.13.1
for the testing infrastructure to prevent the following
bug which is causing gate failures:

https://github.com/docker/docker/issues/28336

Change-Id: I0d3569b589ee792c9aa6eaa7c539219fcc784271