Remove global privilege escalation for deploy_aio.sh
At the moment, deploy_aio.sh is called with sudo which results
in the environment being wiped.  This means that all Zuul
environment variables are removed and therefore when zuul-cloner
runs to checkout the branch of kolla-ansible, it will always
fallback to master as ZUUL_BRANCH is unset.

This patch removes the global usage of sudo in the tox.ini file
as there is existing usage of sudo in the bash script itself. This
will mean that we will only escalate privileges when needed rather
for the whole script and the environment variables should be passed
on properly resulting in the matching release of kolla-ansible being
checked out.

Closes-Bug: #1673756

Change-Id: I477a843e696be7136c020be6408afea01967879d