Switch to Kolla virtual environment for Bifrost
By default, the Bifrost Ansible playbooks will naturally run
everything against /usr/bin/python which means that some modules,
such as Shade when it has to be installed will use the system
Python installation.

By adding `ansible_python_interpreter` and pointing it towards
our Python interpreter, it fixes the failures related to system
Python and conflicts related to it (which in this case, resulted
in Shade failing to build properly).

Closes-Bug: #1673776

Change-Id: I2a5a2a3e2627a313b3306aeb1a910efe0e209496