Turn on SSH for Keystone for all distros.
With the previous patch that was done to enable SSH access, it was
mistakingly done to be handled only in the Ubuntu case.  It should
always be done as SSH is always needed.

This is only a problem for the backport, the change introduced in
master and stable/ocata relies on the new configure_user macro
which runs regardless.

Change-Id: Ibef580748cc139f0ebf207609f934f1469222624
Closes-Bug: #1635795