Gate fixes: disable bifrost-base, fix ubuntu base image
At the moment, Bifrost has caused a few failures which have
blocked the gate, the latest of which is documented in the
bug listed below (alongside with another one that caused another
gate failure).

In order to let patches flow though, we're disabling it
from the gate build checks as agreed on the Kolla meeting on
the 22nd of March 2017.  Also, the Ubuntu base image seems
to have a missing /etc/protocols which is created by the
netbase package.  This has been added to the base images.

There are issues with subunit getting a very large log all
at once which causes failures that is causing issues.  We
also enable logging to a local folder which will make
troubleshooting image builds easier to read and avoid
sending a large blob of data to subunit.

Change-Id: Iee1f8b06ec1066015ea5c6a6a245723402e35a6a
Related-Bug: #1674483
Related-Bug: #1673776
Closes-Bug: #1675101