Prevent apache from listening on port 80
A few new images got apache as part of blueprint
apache-packages-for-apis without restricting it to listen on port 80.
This commit fixes it for all images with the exception of heat-base so
that it can be cleanly backported to stable/ocata.

Co-Authored-By: Juan Antonio Osorio Robles <jaosorior@redhat.com>
Change-Id: Idd3bbeffea9233eaed420798e22ea9a3a517c166
Closes-Bug: #1675308
(cherry picked from commit f88ca0f862da4b040b9e0aaac7ac1b952bd0c08b)