Add missing cleanup step for new apache images
A few new images got apache as part of blueprint
apache-packages-for-apis without a cleaning up. Without this step the
containers may fail to restart due to runtime files already present.

This commit fixes it for all images with the exception of heat-base so
that it can be cleanly backported to stable/ocata.

It also removes the unused APACHE_DIR variable from freezer-api and
keystone extend_start.sh script.

Change-Id: I47d75221b3d1aae91f11d3e0da83e221a00315b7
Closes-Bug: #1679565