Move shebang customization into jinja2 block
By setting up the shebang inside a jinja2 block, this allows the
operator to use a different init system of simply disable it. This is
especially useful with docker 1.13 that comes with its own stripped
down init system (a build of tini) that would cause the containers to
run double init systems if used with the '--init' command line
argument.

Change-Id: I220d8cfc81c25471e635b2f5790759f21e788e64
Closes-Bug: #1680770