Skip [plugins,additions]-archive for binary builts
Skip the creation of `plugins-archive` or `additions-archive` for
binary builds as they are only checked-out as part of the source build
process. Doing so also has the nice side-effect of failing the build in
case someone accidently includes an 'ADD plugins-archive /' instruction
to the binary section of a Dockerfile.

Change-Id: I7442fcd9677466dd0878e124402b4d4170a1ab68
Related-Bug: #1681345