Switch to RDO proxy mirrors
We download a large number of packages from trunk.rdoproject.org, the
OpenStack Infra team has setup a per-region caching proxy to speed
up builds and reduce reliance on the internet.

This patch switches builds to use the new mirror and also stops
defining our own mirror variables and rely on the one provided by
nodepool instead.  As part of this patch, we also drop getting the
delorean.repo file from buildlogs.centos.org as it could be out of
sync and essentially broken (until the job to sync runs).  We prefer
going to trunk.rdoproject.org as it will be proxied and cached.

Change-Id: Ia28225de715089bbf5b023d1e134f2fb7da93fbd