Ensure tests run in isolation
Previously, the images were copied locally in the tests as an attempt
to ensure test isolation but the parent image was still pointing to the
global object. This commit updates the images to point to the local
copies.

Rename the images to have more descriptive names and fix an issue in
test_skip_parents where the regex parameter was passed as a string
while it should be an array, that appeared once the images had their
correct parents.

Also bypass docker client memoization to further minimize the risk of
breaking test isolation.

Change-Id: Id7df645284f260cd615188db2efe889c9cf8115d
Closes-Bug: #1683314