Add Apache packages to zaqar container
Some projects that use Kolla docker containers (TripleO) run zaqar
under httpd. This patch adds the httpd package so that it exists in the
container and can optionally by used via Docker entrypoints.

As many of these projects do not require the use of Apache (it is
opt-in) at this point not all deployment frameworks support the use of
it so the existing configuration defaults have been left as-is for now.

Change-Id: I35cfd1c2320eb972890b44668c8f9f0a047a65dc
Partially-Implements: blueprint apache-packages-for-apis