dragonflow: install proper package on all distros
You should NEVER install library packages as 99.99% time is not what you
really need. NEVER!

What was happening? During build of 'pyzmq' it was noticing that there
are no development headers for libzmq and after 10s delay (with huge
warning attached) it was starting to build own 'bundled' copy of libzmq.

Bumped libzmq to newer version as well (in Debian/Ubuntu).

Change-Id: I8bc95cc9c90cc9f4df911a40392b85c9b5dd1261