Only copy symlinks when merge and preserve_properties are set
The kolla-kubernetes config files live in a ..data directory in the
config volume and are symlinked with relative path at the root of the
config volume. This naturally breaks with the link copy feature
introduce in I92a404b8d301937dc1847d42576e30db43951990.

This commit restores the original behavior to copy the target of the
link unless both 'merge' and 'preserve_properties' options are set, in
which case the link itself is copied.

Change-Id: I5d768687208205fb50ef5ab93675f47e98ea6351