Add libvirt group to base image
The libvirt-daemon controls socket access with the libvirt group,
which is created dynamically.  If different containers would get
different GIDs for the libvirt group that would break cross-container
socket access.  Because creation of only a group is not supported and
for the uids and gid to stay in sync, an unused libvirt user is
created as well.

Change-Id: Ie484a6ad1493f4b867c9577fc0beeb8b2e16c371