base: add support for fetching APT keys from files
For RPM distros we use GPG keys as URLs because rpm can handle it.

For Debian based distros we fetch GPG keys from ubuntu keyserver. But it
does not work with OBS repositories where key is provided as file in
repo.

This change adds support for fetching GPG keys for APT use.

Change-Id: I6b205a2bb61207d1549222ba5501536b2743df77