Fix Horizon asset compression not happening on redeploy
A recent change to kolla-ansible (a729dcc4) bind mounts /tmp from the
host to handle uploading of larger images. Unfortunately this is also
where we keep the state file for horizon which controls whether assets
need to be recompressed (.local_settings.md5sum.txt). This means it is
left over on destroy and causes issues with CSS / assets on redeploy.

Change the location of this file to
/var/lib/kolla/.local_settings.md5sum.txt so it is cleaned up correctly
along with the container.

Change-Id: Icfc71266cabedc2c1a3d30590cada3f774bc0d2c
Closes-Bug: #1721281
(cherry picked from commit cad0a1e697ebc2caf9dd9dc6e4abe6ccaccf1b07)